llm/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
langchain==0.1.19
langchain-community==0.0.38
langserve==0.3.1
python-dotenv==1.0.1
httpx>=0.25.0
pydantic>=2.0.0